smart-knowledge-agent/requirements.txt
python-dotenv==1.0.1
langchain==0.2.16
langchain-openai==0.1.21
faiss-cpu==1.8.0.post1
chromadb==0.5.5
pypdf==5.0.1
markdown==3.6
unstructured==0.15.7
tiktoken==0.7.0
openai==1.45.0
streamlit==1.37.1
numpy==1.26.4
pandas==2.2.2
uvicorn==0.30.6
fastapi==0.114.0
